FIX Selected tag height. Move Readonly to own class.
