+++ css/TagField.css
@@ -50,7 +50,7 @@
 }
 
 .select2-selection__choice {
-    height: 13px;
+    height: 22px;
     line-height: 13px;
     -webkit-border-radius: 3px;
     -moz-border-radius: 3px;
