+++ src/.upgrade.yml
@@ -1,4 +1,4 @@
 mappings:
   StringTagField: SilverStripe\TagField\StringTagField
   TagField: SilverStripe\TagField\TagField
-  TagField_Readonly: SilverStripe\TagField\TagField_Readonly
+  TagField_Readonly: SilverStripe\TagField\TagField\Readonly
